% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_slide_graph}
\alias{draw_slide_graph}
\title{Plot SRT Slide}
\usage{
draw_slide_graph(
  meta_data,
  edge_df = NULL,
  threshold = NULL,
  col_sel,
  pal = NULL,
  flip = T
)
}
\arguments{
\item{meta_data}{A Data Frame contains at least coordinates}

\item{edge_df}{A Data Frame contains the edges information, at least 5 columns: \verb{x, y, xend, yend, weight}}

\item{threshold}{Threshold of edge weight for displaying edges}

\item{col_sel}{column name in \code{meta_data}}

\item{pal}{Palette}

\item{flip}{Whether swap x-y coordinates}
}
\value{
A Heatmap-class object.
}
\description{
Draw SRT spot array with customized labels
}
\examples{
#TBD
}
