% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init}
\alias{init}
\title{Package Initialization}
\usage{
init(verbose = F)
}
\arguments{
\item{verbose}{Show initialization information}
}
\value{
Void
}
\description{
Load Required package
}
\examples{
init()
}
