% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{louvain_w_cor}
\alias{louvain_w_cor}
\title{Louvain with Correlation Matrix}
\usage{
louvain_w_cor(cor_mat_, nn_ = 10, res_ = 1)
}
\arguments{
\item{cor_mat_}{Correlation Matrix}

\item{nn_}{?}

\item{res_}{?}
}
\value{
Clustered Data
}
\description{
Louvain with Correlation Matrix
}
\examples{
No Example
}
