% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myPreprocess}
\alias{myPreprocess}
\title{Preprocess}
\usage{
myPreprocess(data_mat, scale_coef = 10000)
}
\arguments{
\item{data_mat}{Expression Matrix}

\item{scale_coef}{scale factor}
}
\value{
Void
}
\description{
Do column scale
}
\examples{
# No Example
}
