% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rasterSeuratList}
\alias{rasterSeuratList}
\title{Rasterize Single cell datasets}
\usage{
rasterSeuratList(seu_ls, preserveAnno = NULL, tile_width = 3, verbose = F)
}
\arguments{
\item{spe}{Spatial Experiment or Singel Cell Experiment Object}

\item{sample_id}{Column name contain sample IDs}

\item{sel_assay}{Select assay to}

\item{sel_col}{Columns to be included in the \code{meta.data}}

\item{col_name}{Names of the selected columns}
}
\value{
A list of Seurat Objects
}
\description{
Convert spatialExperiment object to a list of Seurat objects based on sample id
}
\examples{
No Example
}
