% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spe2SeuList}
\alias{spe2SeuList}
\title{Convert spe to seurat list}
\usage{
spe2SeuList(
  spe,
  sel_assay = "logcounts",
  sample_id = "sample_id",
  sel_col = c("layer_guess_reordered_short", "spatialLIBD"),
  col_name = c("layer", "spatialLIBD")
)
}
\arguments{
\item{spe}{Spatial Experiment or Singel Cell Experiment Object}

\item{sel_assay}{Select assay to}

\item{sample_id}{Column name contain sample IDs}

\item{sel_col}{Columns to be included in the \code{meta.data}}

\item{col_name}{Names of the selected columns}
}
\value{
A list of Seurat Objects
}
\description{
Convert spatialExperiment object to a list of Seurat objects based on sample id
}
