% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stage_1}
\alias{stage_1}
\title{Niche Identification}
\usage{
stage_1(
  seu_ls,
  cor_threshold = 0.2,
  nn = 12,
  nn_2 = 20,
  cl_resolution = 10,
  top_pcs = 30,
  cl_min = 5,
  find_HVG = T,
  hvg = 2000,
  cor_met = "PC",
  edge_smoothing = T,
  use_glmpca = T,
  verbose = F
)
}
\arguments{
\item{seu_ls}{A list of Seurat Objects, columns \verb{coord_x,coord_y} must be included in the \code{meta.data}}

\item{cor_threshold}{Threshold for edge pruning}

\item{nn}{Number of NN for graph building}

\item{nn_2}{For SNN correlation only}

\item{cl_resolution}{Louvain resolution for clustering}

\item{top_pcs}{Number of selected top PCs}

\item{cl_min}{Minimal nodes in identified niches}

\item{find_HVG}{Whether find HVG for downstream analysis or use full gene list}

\item{hvg}{HVG numbers, default: \code{2000}}

\item{cor_met}{Correlation calculation method, available methods:\verb{("PC","HVG")}}

\item{edge_smoothing}{Perform smoothed edge detection}

\item{use_glmpca}{Use GLMPCA or regular PCA}

\item{verbose}{Output clustering information}
}
\value{
A list of Seurat Objects
}
\description{
Main function
}
\examples{
# TBD
}
