% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stage_2}
\alias{stage_2}
\title{Integrative Clustering}
\usage{
stage_2(
  seu_ls,
  top_pcs = 30,
  nn_2 = 10,
  cl_key = "merged_cluster",
  rtn_seurat = F,
  method = "louvain",
  hly_cor = 0.9,
  hvg = 2000,
  cor_met = "PC",
  use_glmpca = F,
  resolution = 1,
  rare_ct = "none",
  verbose = F
)
}
\arguments{
\item{seu_ls}{A list of Seurat Objects from \code{stage_1} output}

\item{top_pcs}{Number of selected top PCs}

\item{nn_2}{MNN parameter}

\item{cl_key}{Column contains primary cluster labels}

\item{rtn_seurat}{Return a Seurat object with averaged expression profile and integrative clustering results}

\item{method}{Secondary Clustering algorithm, available methods:\verb{("Louvain","MNN")}}

\item{hly_cor}{Threshold for highly correlated clusters}

\item{hvg}{HVG number for secondary clustering}

\item{cor_met}{Correlation calculation method, available methods:\verb{("PC","HVG")}}

\item{use_glmpca}{Use GLMPCA or regular PCA}

\item{resolution}{Secondary clustering resolution}

\item{rare_ct}{Rare cell type detection method, available methods:\verb{("a","m","none")}}

\item{verbose}{Output clustering information}
}
\value{
A list of Seurat Objects
}
\description{
Main function 2
}
\examples{
# TBD
}
